+++ test/unit.js
@@ -7,7 +7,7 @@ import './unit/api/fragment';
 import './unit/api/init';
 import './unit/api/properties';
 import './unit/api/ready';
-import './unit/api/render';
+import './unit/api/state';
 import './unit/constructor';
 import './unit/document-observer';
 import './unit/dom';